netcdf: patch to fix hdf5 1.8.13 missing symbols
HDF5 1.8.13 removes symbols related to MPI POSIX VFD, leading to
errors when linking hdf5 and netcdf5 such as "undefined reference to
`_H5Pset_fapl_mpiposix’".

This commit adds a patch to netcdf that fixes these errors, and fixes
issue #911. The patch has already been added to netcdf, and will be
included in the next release.