elemental: fix build errors in {stable, devel}
* remove conflicting CMake macros in {stable, devel}

Fix build errors for stable (0.85) & devel (0.86-rc1) versions of
Elemental. Elemental rolled its own CMake macros in stable & devel;
these features were then incorporated in a recent version of CMake
(not sure which one). Consequently, the Elemental versions of CMake
macros now conflict with the built-in CMake macros. This patch removes
the Elemental versions of the conflicting macros, and is ported from
commit 96fac79ce19bb6a917c49c27b31a3099b88b4652 of the Elemental
repository. See the e-mail discussion at
http://permalink.gmane.org/gmane.comp.mathematics.elemental.maint/138
for supporting information.

* elemental: fix incorrect symlinking

Change the symlink commands from `ln_sf` to the `symlink` methods of
`lib` and `include`. Correct the symlinks so that all files in
`libexec/"include"` and `libexec/"lib"` directories are symlinked to
the `include` and `lib` directories with the same base filenames.

* fix check & test commands, move all to brew-test

Elemental's directory structure has been reorganized in its master
branch; the source file corresponding to the AxpyInterface executable
has been moved, and the source file corresponding to the SVD test has
been deleted. These tests have been replaced with tests that should
run on the current stable version (0.85), the development version
(0.86-rc1), and the master branch of the GitHub repository (HEAD). To
conform to Homebrew style guidelines, move these tests into the `brew
test` block. To ensure that the tests are executable after build,
install the examples used in the `brew test` block by default.

* replace the old build-time tests

Remove deprecated check option because it has been updated with
different functionality (build-time tests). Replace the tests that
had been moved to the `brew test` block with Elemental's native
build-time tests. Do not execute these tests by default, because they
can take some time (~15 min on my laptop) to execute.

* copy Make & CMake files for developers

Copy the Make & CMake scripts that include Elemental path information
to enable automatic detection of Elemental compile and link
flags. These files are useful for developers wishing to write software
that depends on Elemental.

Only install CMake config files if installing from HEAD, because these
files do not exist in devel or stable. This line is really a
placeholder for maintainers of the package, because right now, HEAD
does not build, and probably will not build until version 0.87 is
released, due to issues with `metis` and `parmetis` dependencies.

* fix python bindings install for devel

Python bindings were installed in the wrong location, and not vendored
into a Python install via a pth file. Fix this issue by installing the
Python bindings so that they are vendored into the correct location
via a pth file.

For Python bindings in devel, default to the system Python unless the
Mac OS version is sufficiently old.

* fix brew audit --online --strict errors

Update the Elemental formula for style, so that it passes `brew audit`
checks. This change includes moving the current build-time tests to
the `brew test` block.